Fix pattern match for api_url match
The pattern for pulling api_host out of api_url is incorrect, and fails if the URL is just a "http" protocol.  I have left the character class (i.e. [s]) in place as it is a visual clue that it is just the single character, where as the correct substring would either give some confusion (a lone 's?') or more characters ('(?:s)') and still be messy.